requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncer==0.0.8
attrs==25.1.0
backoff==2.2.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
cryptography==43.0.3
datasets==3.2.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
dspy==2.6.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-sso==0.16.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
groq==0.18.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
identify==2.6.6
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.35.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.57.4
magicattr==0.1.6
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
multiprocess==0.70.16
nodeenv==1.9.1
numpy==2.2.2
oauthlib==3.2.2
openai==1.61.0
optuna==4.2.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.1.0
propcache==0.2.1
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
rq==2.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.22.0
virtualenv==20.29.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
